small chanhes

+++ data_gen/qm_calcs/calculateGeomWithQM.py
@@ -3,6 +3,7 @@ from ase.io import read, write
 import os, shutil
 #  from dftd4 import D4_model
 from ase.calculators.orca import ORCA
+from ase.calculators.orca import OrcaProfile
 #from gpaw import GPAW, PW
 
 #  import numpy as np
@@ -19,7 +20,7 @@ from orca_io import (read_orca_h_charges,
 
 
 
-def prepareDDECinput(label):
+def prepareDDECinput(base):
     input_txt = f"""
         <periodicity along A, B, and C vectors>
         .false.
@@ -32,7 +33,7 @@ def prepareDDECinput(label):
         </atomic densities directory complete path>
 
         <input filename>
-        {label}.wfx
+        {base}.wfx
         </input filename>
 
         <charge type>
@@ -47,12 +48,12 @@ def prepareDDECinput(label):
         print(input_txt, file=fl)
 
 
-def orca_calculator(label, calc_type,  n_task, initial_gbw=['', '']):
-    return ORCA(
-        label=label,
-        maxiter=250,
+def orca_calculator(orca_path, base, calc_type,  n_task, initial_gbw=['', '']):
+    calc = ORCA(
+        profile=OrcaProfile(orca_path),
+        # label=label,
         charge=0, mult=1,
-        orcasimpleinput= calc_type.upper() \
+        orcasimpleinput=f"{calc_type.upper()}" \
             + ' PBE D4 DEF2-TZVP DEF2/J RIJDX MINIPRINT NOPRINTMOS NoKeepInts NOKEEPDENS CHELPG AIM ' \
             + initial_gbw[0],
         orcablocks= '%scf Convergence tight \n maxiter 250 end \n %output \n' \
@@ -60,19 +61,24 @@ def orca_calculator(label, calc_type,  n_task, initial_gbw=['', '']):
             + str(n_task) + ' end' \
             + initial_gbw[1]
     )
+    return calc
 
 
 class CaculateData():
-    def __init__(self, properties, calc_type, n_task, in_extxyz_path, out_extxyz_path, csv_path):
+    def __init__(self, orca_path, calc_type, n_task, in_extxyz_path, out_extxyz_path, csv_path, base_dir=None):
 
+        self.orca_path = orca_path
         self.in_extxyz_path = in_extxyz_path
         self.out_extxyz_path = out_extxyz_path
         self.csv_path = csv_path
+
+        self.BASE_DIR = base_dir
+        if base_dir is None:
+            self._setBASE_DIR()
         #  self.i = 0
 
         self.atoms_list = None
         self._loadAtaoms()
-        self.properties = properties
 
         self.calc_type = calc_type
         self.n_task = n_task
@@ -80,6 +86,12 @@ class CaculateData():
         self._checkCSVFile()
 
         self.rm_files = False
+
+    def _setBASE_DIR(self):
+        self.BASE_DIR = os.getcwd()
+
+    def _goBASE_DIR(self):
+        os.chdir(self.BASE_DIR)
 
     def _loadAtaoms(self):
         self.atoms_list = read(self.in_extxyz_path, index=":")
@@ -100,6 +112,10 @@ class CaculateData():
         df_calculated_files_new.to_csv(self.csv_path, mode='a', header=False, index=None)
 
     def _calculate_data(self, idx):
+
+        # to prevent nested out dir for same time proccessors
+        self._goBASE_DIR()
+
         atoms = self.atoms_list[idx]
         try:
             label = atoms.info["label"]
@@ -120,26 +136,26 @@ class CaculateData():
 
         #  label = "orca_%s" %label
 
-        cwd = os.getcwd()
-
         #  full path
-        OUT_DIR = Path(cwd) / Path("run_" + f"{self.calc_type}_" + self.in_extxyz_path.split('/')[-1].split(".")[0]) / Path(label)
+        OUT_DIR = Path(self.BASE_DIR) / Path(f"run_{self.calc_type}_" + self.in_extxyz_path.split('/')[-1].split(".")[0]) / Path(label)
         OUT_DIR.mkdir(parents=True, exist_ok=True)
 
-        GBW_DIR = Path(cwd) / Path("run_" + self.in_extxyz_path.split('/')[-1].split(".")[0])
+        GBW_DIR = Path(self.BASE_DIR) / Path(f"run_{self.calc_type}_" + self.in_extxyz_path.split('/')[-1].split(".")[0])
         initial_gbw_name = "initial_" + label.split("_")[0] + ".gbw"
         initial_gbw_file = [flname for flname in os.listdir(GBW_DIR) if ".gbw" in flname]
 
         print(f"working in {OUT_DIR} directory\n")
         os.chdir(OUT_DIR)
 
+        # TODO: to be fixed
+        label = "orca"
         try:
             if len(initial_gbw_file) == 1:
                 shutil.copy2(f"{GBW_DIR}/{initial_gbw_name}", OUT_DIR)
                 initial_gbw = ['MORead',  '\n%moinp "{}"'.format(initial_gbw_name)]
-                atoms.set_calculator(orca_calculator(label, self.calc_type, self.n_task, initial_gbw))
+                atoms.set_calculator(orca_calculator(self.orca_path, label, self.calc_type, self.n_task, initial_gbw))
             else:
-                atoms.set_calculator(orca_calculator(label, self.calc_type, self.n_task))
+                atoms.set_calculator(orca_calculator(self.orca_path, label, self.calc_type, self.n_task))
 
             # orca calculation start
             atoms.get_potential_energy()
@@ -162,13 +178,11 @@ class CaculateData():
                 os.system("mv %s.gbw %s/%s" %(label, GBW_DIR, initial_gbw_name))
 
             #  os.system("rm %s*" %label)
-            os.chdir(cwd)
+            os.chdir(self.BASE_DIR)
             write(self.out_extxyz_path, atoms, append=True)
             shutil.rmtree(OUT_DIR)
-                #  self.i += 1
         except:
-            #  print("Error for %s" %label)
-            #  self.i += 1
+            print("Error for %s" %label)
 
             # remove this non SCF converged file from xyz directory.
             if self.rm_files:
@@ -177,8 +191,9 @@ class CaculateData():
 
             # remove all orca temp out files related to label from runGeom directory.
             #  os.system("rm %s/%s*" %(OUT_DIR, label))
-            shutil.rmtree(OUT_DIR)
-            os.chdir(cwd)
+            if self.calc_type.lower() == "sp":
+                shutil.rmtree(OUT_DIR)
+            os.chdir(self.BASE_DIR)
             return None
 
     def countAtoms(self):
